refactor: rename ConfigLoader and ModelLoader files, update imports, and adjust ChatsConfig structure
